web/requirements.txt
Django==3.0.5
psycopg2-binary==2.8.4
uwsgi==2.0.18
djangorestframework==3.11.0
django-cors-headers==3.2.1
djangorestframework-simplejwt==4.4.0